requirements.txt
elasticsearch==8.13.0
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.0
requests==2.31.0
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
imbalanced-learn==0.12.0
python-dotenv==1.0.0

# PHASE 0: Encryption
cryptography==42.0.0

# PHASE 1: OAuth2/JWT Authentication 
PyJWT==2.8.1
passlib==1.7.4
python-multipart==0.0.6
slowapi==0.1.9

# API & Backend
fastapi==0.104.1
uvicorn==0.24.0

# Elasticsearch & Monitoring
elasticsearch==8.13.0
